Control de borrado de usuarios
No se estaba controlando del lado del servidor que un administrador no pueda borrar a otro administrador

+++ app/Http/Controllers/UserController.php
@@ -82,6 +82,9 @@ class UserController extends Controller
     public function destroy($userName)
     {   
         $user = User::where('user_name', $userName)->first();
+        if($user->role == 'administrator'){
+            abort(403);
+        }
         $user->delete();
         return redirect('/users')->with('success', 'User '.$userName.' was deleted successfully!');
     }
